+++ HISTORY.rst
@@ -6,10 +6,13 @@ History
 x.x.x
 ~~~~~
 
-- Update CDN link to the widget
+- Add file.create_local_copy and file.create_remote_copy methods.
+- Add new ``make_public`` and ``pattern`` parameters to file.create_remote_copy method.
+- Add new ``store`` parameter to file.create_local_copy methods.
+- Update CDN link to the widget.
 - Use wildcard ``2.x`` to always get the latest
-  patch or minor version of the ``2`` major version of the widget,
-  see `widget changelog`_  now and then :)
+  patch or minor version of widget version ``2``. 
+  See `widget changelog` (see `widget changelog`_).
 
 2.1.0
 ~~~~~